+++ .github/CODEOWNERS
@@ -163,8 +163,16 @@
 /tensorflow_addons/optimizers/yogi.py @manzilz
 /tensorflow_addons/optimizers/tests/yogi_test.py @manzilz
 
-/tensorflow_addons/rnn/cell.py @qlzh727 @pedrolarben
-/tensorflow_addons/rnn/tests/cell_test.py @qlzh727 @pedrolarben
+/tensorflow_addons/rnn/esn_cell.py @qlzh727 @pedrolarben
+/tensorflow_addons/rnn/tests/esn_cell_test.py @qlzh727 @pedrolarben
+/tensorflow_addons/rnn/layer_norm_lstm_cell.py @qlzh727 @pedrolarben
+/tensorflow_addons/rnn/tests/layer_norm_lstm_cell_test.py @qlzh727 @pedrolarben
+/tensorflow_addons/rnn/layer_norm_simple_rnn_cell.py @qlzh727 @pedrolarben
+/tensorflow_addons/rnn/tests/layer_norm_simple_rnn_cell_test.py @qlzh727 @pedrolarben
+/tensorflow_addons/rnn/nas_cell.py @qlzh727 @pedrolarben
+/tensorflow_addons/rnn/tests/nas_cell_test.py @qlzh727 @pedrolarben
+/tensorflow_addons/rnn/peephole_lstm_cell.py @qlzh727 @pedrolarben
+/tensorflow_addons/rnn/tests/peephole_lstm_cell_test.py @qlzh727 @pedrolarben
 
 /tensorflow_addons/seq2seq/ @qlzh727 @guillaumekln
